+++ configs/eslint-config-react/src/index.ts
@@ -40,7 +40,6 @@ const config: ESLint.ConfigData = {
     ],
 
     "react/jsx-uses-react": "error",
-    "react/react-in-jsx-scope": "error",
     "react/jsx-uses-vars": "error",
     "react/jsx-key": "error",
     "react/display-name": ["error", { ignoreTranspilerName: false }],